Breakout the tests into their own files
